feat: eslint support ts
