+++ packages/fiddle/package.json
@@ -37,6 +37,7 @@
   },
   "devDependencies": {
     "@types/lz-string": "^1.3.34",
+    "@typescript-eslint/typescript-estree": "^5.43.0",
     "raw-loader": "^4.0.2"
   },
   "scripts": {
